refactor: 1115 `Hash` implementations moved in `vcdm`
