+++ packages/core/src/certificate/certificate.ts
@@ -1,6 +1,6 @@
 import { CertificateSignature } from '@vechain/sdk-errors';
 import fastJsonStableStringify from 'fast-json-stable-stringify';
-import { Blake2b256 } from '../hash';
+import { Blake2b256 } from '../vcdm/hash/Blake2b256';
 import { secp256k1 } from '../secp256k1';
 import { Address, Hex, Txt } from '../vcdm';
 import { type Certificate } from './types';
