+++ packages/core/src/vcdm/hash/Blake2b256.ts
@@ -1,7 +1,7 @@
 import { blake2b as nh_blake2b } from '@noble/hashes/blake2b';
 import { InvalidOperation } from '@vechain/sdk-errors';
-import { Hex, Txt, type Hash } from '../vcdm';
-import { HexUInt } from '../vcdm/HexUInt';
+import { Hex, Txt, type Hash } from '../index';
+import { HexUInt } from '../HexUInt';
 
 /**
  * Represents the result of an [BLAKE](https://en.wikipedia.org/wiki/BLAKE_(hash_function)) [BlAKE2B 256](https://www.blake2.net/) hash operation.
