+++ packages/core/src/vcdm/hash/Keccak256.ts
@@ -1,7 +1,7 @@
 import { keccak_256 as nh_keccak_256 } from '@noble/hashes/sha3';
 import { InvalidOperation } from '@vechain/sdk-errors';
-import { Hex, Txt, type Hash } from '../vcdm';
-import { HexUInt } from '../vcdm/HexUInt';
+import { Hex, Txt, type Hash } from '../index';
+import { HexUInt } from '../HexUInt';
 
 /**
  * Represents the result of an [SHA-3](https://en.wikipedia.org/wiki/SHA-3) [KECCAK 256](https://keccak.team/keccak.html) hash operation.
