+++ packages/core/src/vcdm/hash/index.ts
@@ -1,3 +1,4 @@
 export * from './Blake2b256';
+export * from './Hash';
 export * from './Keccak256';
 export * from './Sha256';
