+++ packages/core/tests/vcdm/hash/fixture.ts
@@ -1,4 +1,4 @@
-import { Hex, Txt } from '../../src';
+import { Hex, Txt } from '../../../src';
 
 const CONTENT = Hex.of(
     Txt.of('Hello world - Здравствуйте - こんにちは!').bytes
